Fix aitalk no longer firing at all
With the current version, aitalk will never fire, because OnSecondChange is used to control aitalk and therefore _is_second_change is always true when it comes to aitalk.
I've reverted the check for the _event variable; now it once again checks to see if the event has changed from OnSecondChange to OnAITalkNewEvent. This seems to have fixed all of the bugs on my end.